+++ src/main/java/com/jwplayer/jwplatform/client/PlayerBiddingConfigClient.java
@@ -10,30 +10,30 @@ import com.jwplayer.jwplatform.exception.JWPlatformException;
 import com.jwplayer.jwplatform.rest.HttpCalls;
 
 /**
- * JW Platform VPBConfigs API client.
+ * JW Platform PlayerBiddingConfigs API client.
  *
  * <p>
- * An API client for the JW Platform VPBConfigs API. For the API documentation
+ * An API client for the JW Platform PlayerBiddingConfigs API. For the API documentation
  * see: <a href=
  * "https://developer.jwplayer.com/jwplayer/reference#introduction-to-api-v2">Introduction
  * to api v2</a>
  *
  * <p>
- * Example: VPBConfigsClient client = VPBConfigsClient.getClient(secret);
+ * Example: PlayerBiddingConfigClient client = PlayerBiddingConfigClient.getClient(secret);
  */
-public class VPBConfigsClient extends JWPlatformClientV2 {
+public class PlayerBiddingConfigClient extends JWPlatformClientV2 {
 
 	private String path;
 	private final String secret;
 
 	/**
-	 * Instantiate a new {@code VPBConfigsClient} instance.
+	 * Instantiate a new {@code PlayerBiddingConfigClient} instance.
 	 *
 	 * @param secret - your api secret
 	 */
-	private VPBConfigsClient(String secret) {
+	private PlayerBiddingConfigClient(String secret) {
 		this.secret = secret;
-		this.path = "https://api.jwplayer.com/v2/sites/%s/advertising/vpb_configs/";
+		this.path = "https://api.jwplayer.com/v2/sites/%s/advertising/player_bidding_configs/";
 		headers = new HashMap<>();
 		headers.put("Authorization", "Bearer " + this.secret);
 		headers.put("accept", "application/json");
@@ -41,11 +41,11 @@ public class VPBConfigsClient extends JWPlatformClientV2 {
 	}
 
 	/**
-	 * see {@link #VPBConfigsClient(String)}.
+	 * see {@link #PlayerBiddingConfigClient(String)}.
 	 */
-	public static VPBConfigsClient getClient(String secret) {
+	public static PlayerBiddingConfigClient getClient(String secret) {
 		Preconditions.checkNotNull(secret, "API Secret must not be null!");
-		return new VPBConfigsClient(secret);
+		return new PlayerBiddingConfigClient(secret);
 	}
 
 	/**
@@ -54,8 +54,7 @@ public class VPBConfigsClient extends JWPlatformClientV2 {
 	 * @param params - Parameters to be included in the request
 	 * @return JSON response from Media API
 	 * @throws JWPlatformException See <a href=
-	 *                             "https://developer.jwplayer.com/jwplayer/reference?showHidden=93052#get_v2-sites-site-id-advertising-vpb-configs">List
-	 *                             configs</a>
+	 *                             "https://developer.jwplayer.com/jwplayer/reference/get_v2-sites-site-id-advertising-player-bidding-configs">List Configs</a>
 	 */
 	public JSONObject listConfigs(String siteId, Map<String, String> params) throws JWPlatformException {
 		Preconditions.checkNotNull(siteId, "Site ID must not be null!");
@@ -69,8 +68,7 @@ public class VPBConfigsClient extends JWPlatformClientV2 {
 	 * @param bodyParams - Parameters to be included in the request body
 	 * @return JSON response from Media API
 	 * @throws JWPlatformException See <a href=
-	 *                             "https://developer.jwplayer.com/jwplayer/reference?showHidden=93052#post_v2-sites-site-id-advertising-vpb-configs">Create
-	 *                             config</a>
+	 *                             "https://developer.jwplayer.com/jwplayer/reference/post_v2-sites-site-id-advertising-player-bidding-configs">Create Configs</a>
 	 */
 	public JSONObject createConfig(String siteId, Map<String, String> bodyParams) throws JWPlatformException {
 		Preconditions.checkNotNull(siteId, "Site ID must not be null!");
@@ -86,8 +84,7 @@ public class VPBConfigsClient extends JWPlatformClientV2 {
 	 * @param bodyParams - Parameters to be included in the request body
 	 * @return JSON response from Media API
 	 * @throws JWPlatformException See <a href=
-	 *                             "https://developer.jwplayer.com/jwplayer/reference?showHidden=93052#patch_v2-sites-site-id-advertising-vpb-configs-config-id-">Update
-	 *                             config</a>
+	 *                             "https://developer.jwplayer.com/jwplayer/reference/patch_v2-sites-site-id-advertising-player-bidding-configs-config-id-">Update Config</a>
 	 */
 	public JSONObject updateConfig(String siteId, String configId, Map<String, String> bodyParams)
 			throws JWPlatformException {
@@ -105,8 +102,7 @@ public class VPBConfigsClient extends JWPlatformClientV2 {
 	 * @param params   - Parameters to be included in the request
 	 * @return JSON response from Media API
 	 * @throws JWPlatformException See <a href=
-	 *                             "https://developer.jwplayer.com/jwplayer/reference?showHidden=93052#get_v2-sites-site-id-advertising-vpb-configs-config-id-">Get
-	 *                             config by ID</a>
+	 *                             "https://developer.jwplayer.com/jwplayer/reference/get_v2-sites-site-id-advertising-player-bidding-configs-config-id-">Get Config By ID</a>
 	 */
 	public JSONObject getConfigById(String siteId, String configId, Map<String, String> params)
 			throws JWPlatformException {
@@ -122,8 +118,7 @@ public class VPBConfigsClient extends JWPlatformClientV2 {
 	 * @param configId - Unique identifier for a resource
 	 * @return JSON response from Media API
 	 * @throws JWPlatformException See <a href=
-	 *                             "https://developer.jwplayer.com/jwplayer/reference?showHidden=93052#delete_v2-sites-site-id-advertising-vpb-configs-config-id-">Delete
-	 *                             config</a>
+	 *                             "https://developer.jwplayer.com/jwplayer/reference/delete_v2-sites-site-id-advertising-player-bidding-configs-config-id-">Delete Config</a>
 	 */
 	public JSONObject deleteConfig(String siteId, String configId) throws JWPlatformException {
 		Preconditions.checkNotNull(siteId, "Site ID must not be null!");
@@ -138,12 +133,11 @@ public class VPBConfigsClient extends JWPlatformClientV2 {
 	 * @param bodyParams - Parameters to be included in the request body
 	 * @return JSON response from Media API
 	 * @throws JWPlatformException See <a href=
-	 *                             "https://developer.jwplayer.com/jwplayer/reference?showHidden=93052#put_v2-sites-site-id-advertising-update-schedules-vpb-config">Update
-	 *                             schedules</a>
+	 *                             "https://developer.jwplayer.com/jwplayer/reference/put_v2-sites-site-id-advertising-update-schedules-player-bidding-configs">Update Schedules</a>
 	 */
 	public JSONObject updateSchedules(String siteId, Map<String, String> bodyParams) throws JWPlatformException {
 		Preconditions.checkNotNull(siteId, "Site ID must not be null!");
-		this.path = "https://api.jwplayer.com/v2/sites/" + siteId + "/advertising/update_schedules_vpb_config/";
+		this.path = "https://api.jwplayer.com/v2/sites/" + siteId + "/advertising/update_schedules_player_bidding_configs/";
 		return HttpCalls.request(this.path, bodyParams, false, "PUT", headers);
 	}