+++ src/test/java/com/jwplayer/jwplatform/client/ThumbnailsClientTest.java
@@ -25,13 +25,13 @@ import com.jwplayer.jwplatform.rest.HttpCalls;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({ HttpCalls.class })
-public class VPBConfigsClientTest {
+public class ThumbnailsClientTest {
 
-	VPBConfigsClient vpbConfigsClient = VPBConfigsClient.getClient("fakeSecret");
+	ThumbnailsClient thumbnailClient = ThumbnailsClient.getClient("fakeSecret");
 
 	@Test
 	public void testAllMethods() throws JWPlatformException {
-		vpbConfigsClient.addHeader("test", "testVal");
+		thumbnailClient.addHeader("test", "testVal");
 		mockStatic(HttpCalls.class);
 		when(HttpCalls.request(anyString(), anyMap(), anyBoolean(), eq("GET"), anyMap()))
 				.thenReturn(new JSONObject("{\"code\":\"success\"}"));
@@ -43,31 +43,27 @@ public class VPBConfigsClientTest {
 				.thenReturn(new JSONObject("{\"code\":\"Update successful\"}"));
 		when(HttpCalls.request(anyString(), anyMap(), anyBoolean(), eq("PUT"), anyMap()))
 				.thenReturn(new JSONObject("{\"code\":\"Put successful\"}"));
-		JSONObject listResp = vpbConfigsClient.listConfigs("siteId", new HashMap<>());
+		JSONObject listResp = thumbnailClient.listThumbnails("siteId", new HashMap<>());
 		assertEquals(listResp.get("code"), "success");
-		JSONObject deleteResp = vpbConfigsClient.deleteConfig("siteId", "configId");
+		JSONObject deleteResp = thumbnailClient.deleteThumbnail("siteId", "thumbnailId");
 		assertEquals(deleteResp.get("code"), "Deletion success!");
 		Map<String, String> params = new HashMap<>();
 		params.put("title", "test");
-		vpbConfigsClient.createConfig("siteId", new HashMap<>());
-		vpbConfigsClient.createConfig("siteId", params);
-		vpbConfigsClient.updateConfig("siteId", "configId", params);
-		vpbConfigsClient.updateConfig("siteId", "configId", new HashMap<>());
-		vpbConfigsClient.getConfigById("siteId", "configId", new HashMap<>());
-		JSONObject updateAdResp = vpbConfigsClient.updateSchedules("siteId", params);
-		assertEquals(updateAdResp.get("code"), "Put successful");
-		vpbConfigsClient.updateSchedules("siteId", new HashMap<>());
-		PowerMockito.verifyStatic(HttpCalls.class, Mockito.times(9));
+		thumbnailClient.createThumbnail("siteId", new HashMap<>());
+		thumbnailClient.createThumbnail("siteId", params);
+		thumbnailClient.updateThumbnail("siteId", "thumbnailId", params);
+		thumbnailClient.deleteThumbnail("siteId", "thumbnailId");
+		PowerMockito.verifyStatic(HttpCalls.class, Mockito.times(6));
 		HttpCalls.request(anyString(), anyMap(), anyBoolean(), anyString(), anyMap());
-		vpbConfigsClient.removeHeader("test");
+		thumbnailClient.removeHeader("test");
 	}
 
 	@Test(expected = JWPlatformException.class)
-	public void testGetVpbConfigsException() throws JSONException, JWPlatformException {
-		vpbConfigsClient.addHeader("test", "testVal");
+	public void testGetOriginalsException() throws JSONException, JWPlatformException {
+		thumbnailClient.addHeader("test", "testVal");
 		mockStatic(HttpCalls.class);
 		when(HttpCalls.request(anyString(), anyMap(), anyBoolean(), eq("GET"), anyMap()))
 				.thenThrow(new JWPlatformException("some exception occured"));
-		vpbConfigsClient.listConfigs("siteId", new HashMap<>());
+		thumbnailClient.listThumbnails("siteId", new HashMap<>());
 	}
 }