+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Add V2 Clients for Text tracks, Tags, Usage, Player Bidding Configs, Originals, Media Renditions, Thumbnails.
+* Update Playlists client to include watchlist.
+
 ## 1.0.0
 
 * Add the following V2 clients-